Prevent plugin from including the wrong `functions.php`
Without a prefix, PHP refers to the `include_path`, which
includes the working directory. When called during a unit test,
the working directory will be that of the theme or plugin under
test, not the typical entry path for WordPress. As a result,
the `include( 'functions.php' );` line was prone to include a
theme's `functions.php` prematurely, rather than including the
plugin's file as intended. Prefixing the include path with `./`
would have a similar effect, but there's no downside to being
explicit.